requirements.txt
flask
flask-socketio
eventlet
pymongo
python-dotenv
flask-login